requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.52.2
pinecone-client==5.0.1
firebase-admin==6.6.0
httpx<0.28

